Fixing Gap in Text Scrolling
The patch to make the gap public api take an int was causing a problem in the shader
Now casting to a float before used by shader

+++ dali-toolkit/internal/text/text-scroller.cpp
@@ -197,12 +197,12 @@ TextScrollerPtr TextScroller::New( ScrollerInterface& scrollerInterface )
 void TextScroller::SetGap( int gap )
 {
   DALI_LOG_INFO( gLogFilter, Debug::Verbose, "TextScroller::SetGap gap[%d]\n", gap );
-  mWrapGap = gap;
+  mWrapGap = static_cast<float>(gap);
 }
 
 int TextScroller::GetGap() const
 {
-  return mWrapGap;
+  return static_cast<int>(mWrapGap);
 }
 
 void TextScroller::SetSpeed( int scrollSpeed )
@@ -252,7 +252,7 @@ TextScroller::TextScroller( ScrollerInterface& scrollerInterface ) : mScrollerIn
                             mScrollDeltaIndex( Property::INVALID_INDEX ),
                             mScrollSpeed( MINIMUM_SCROLL_SPEED ),
                             mLoopCount( 1 ),
-                            mWrapGap( 0 )
+                            mWrapGap( 0.0f )
 {
   DALI_LOG_INFO( gLogFilter, Debug::Verbose, "TextScroller Default Constructor\n" );
 }
@@ -286,7 +286,7 @@ void TextScroller::SetParameters( Actor sourceActor, const Size& controlSize, co
 
   mOffscreenCameraActor.SetY( offScreenSize.height * 0.5f );
 
-  DALI_LOG_INFO( gLogFilter, Debug::Verbose, "TextScroller::SetParameters mWrapGap[%d]\n", mWrapGap )
+  DALI_LOG_INFO( gLogFilter, Debug::Verbose, "TextScroller::SetParameters mWrapGap[%f]\n", mWrapGap )
 
   mScrollingTextActor = Actor::New();
   mScrollingTextActor.AddRenderer( renderer );
